+++ dictionaries/kr.json
@@ -12,6 +12,16 @@
     "dont_have_account": "계정이 없으신가요?",
     "logout": "로그아웃"
   },
+  "landing": {
+    "title": "DrawStarter: 당신의 창작을 시작하는 가장 쉬운 방법",
+    "subtitle": "오늘 무엇을 그릴지 막막할 때, 작업 시간을 꾸준히 확보하고 싶을 때, DrawStarter가 여러분의 창작 여정을 함께합니다.",
+    "feature_1_title": "✍️ 랜덤 주제 생성기",
+    "feature_1_description": "영감이 떠오르지 않을 때, 버튼 하나로 다채로운 드로잉 주제를 즉시 제안받을 수 있습니다.",
+    "feature_2_title": "⏱️ 집중 타이머",
+    "feature_2_description": "설정된 시간 동안 온전히 작업에만 집중할 수 있도록 도와주는 타이머입니다.",
+    "feature_3_title": "📈 작업 시간 자동 기록",
+    "feature_3_description": "타이머를 사용한 시간이 자동으로 기록되어, 나의 작업 패턴과 누적 시간을 한눈에 파악할 수 있습니다."
+  },
   "home": {
     "total_accumulated_time": "총 누적 시간",
     "today_work_time": "오늘의 작업 시간",
